+++ conduit-spotify/src/Concerns/AnalyzesArtists.php
@@ -1,0 +1,77 @@
+<?php
+
+namespace Conduit\Spotify\Concerns;
+
+use Conduit\Spotify\Contracts\ApiInterface;
+
+trait AnalyzesArtists
+{
+    public function analyzeArtists(ApiInterface $api): array
+    {
+        $playlists = $api->getUserPlaylists(50);
+        $artistFrequency = [];
+        $artistTracks = [];
+
+        foreach ($playlists as $playlist) {
+            $tracks = $api->getPlaylistTracks($playlist['id']);
+            foreach ($tracks as $track) {
+                if (!isset($track['track']['artists'][0])) continue;
+                
+                $artist = $track['track']['artists'][0]['name'];
+                $trackName = $track['track']['name'];
+                
+                $artistFrequency[$artist] = ($artistFrequency[$artist] ?? 0) + 1;
+                $artistTracks[$artist][] = [
+                    'name' => $trackName,
+                    'playlist' => $playlist['name'],
+                    'uri' => $track['track']['uri'] ?? null,
+                ];
+            }
+        }
+
+        arsort($artistFrequency);
+
+        return [
+            'artist_frequency' => $artistFrequency,
+            'artist_tracks' => $artistTracks,
+            'total_artists' => count($artistFrequency),
+            'most_frequent_artist' => array_key_first($artistFrequency),
+            'most_frequent_count' => reset($artistFrequency),
+        ];
+    }
+
+    public function createTopArtistsPlaylist(ApiInterface $api, array $artistTracks, array $artistFrequency): void
+    {
+        $topArtists = array_slice($artistFrequency, 0, 10, true);
+        $selectedTracks = [];
+
+        foreach ($topArtists as $artist => $count) {
+            if (isset($artistTracks[$artist])) {
+                // Take up to 3 tracks per top artist
+                $tracks = array_slice($artistTracks[$artist], 0, 3);
+                foreach ($tracks as $track) {
+                    if ($track['uri']) {
+                        $selectedTracks[] = $track['uri'];
+                    }
+                }
+            }
+        }
+
+        if (empty($selectedTracks)) {
+            return;
+        }
+
+        // Create playlist
+        $playlist = $api->createPlaylist(
+            'Top Artists Mix',
+            'Curated mix from your most frequent artists. Generated by Conduit Analytics.',
+            false
+        );
+
+        // Add tracks in chunks
+        $chunks = array_chunk($selectedTracks, 100);
+        foreach ($chunks as $chunk) {
+            $api->addTracksToPlaylist($playlist['id'], $chunk);
+        }
+    }
+}
